requirements.txt
aivcontracts==1.0.18
contourpy==1.0.7
cycler==0.11.0
filterpy==1.4.5
fonttools==4.38.0
imageio==2.25.0
kafka-python==2.0.2
kiwisolver==1.4.4
matplotlib==3.6.3
networkx==3.0
numpy==1.24.1
packaging==23.0
pillow==9.4.0
pyparsing==3.0.9
python-configuration==0.8.2
python-dateutil==2.8.2
pywavelets==1.4.1
scikit-image==0.19.3
scipy==1.9.3
shapely==1.8.5.post1
six==1.16.0
tifffile==2023.1.23.1
aiv-event-factory==1.1.2
